move T.A.L. to separate file
